refactor(firestore-translate-text): move package.json into functions dir
