+++ firestore-translate-text/functions/__tests__/mocks/console.ts
