+++ firestore-translate-text/functions/__tests__/mocks/firestore.ts
