+++ firestore-translate-text/functions/__tests__/test.types.d.ts
