+++ firestore-translate-text/functions/__tests__/tsconfig.json
