+++ firestore-translate-text/functions/jest.config.js
