+++ firestore-translate-text/functions/package.json
@@ -1,15 +1,14 @@
 {
-  "name": "firestore-translate-text",
-  "description": "Translate messages into multiple languages, provided by user",
-  "author": "Chris Bianca <chris@csfrequency.com>",
+  "name": "firestore-translate-text-functions",
+  "description": "Firebase Cloud Functions for the Firestore Translate Text Extension",
+  "main": "lib/index.js",
   "license": "Apache-2.0",
-  "main": "functions/lib/index.js",
   "scripts": {
     "build": "npm run clean && npm run compile",
-    "clean": "rimraf functions/lib",
+    "clean": "rimraf lib",
     "compile": "tsc",
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "generate-readme": "firebase ext:info . --markdown > README.md"
+    "test": "jest",
+    "generate-readme": "firebase ext:info .. --markdown > ../README.md"
   },
   "dependencies": {
     "@google-cloud/translate": "^4.0.1",
@@ -21,7 +20,10 @@
     "js-yaml": "^3.13.1",
     "mocked-env": "^1.3.1",
     "rimraf": "^2.6.3",
-    "typescript": "^3.2.4"
+    "typescript": "^3.2.4",
+    "@types/jest": "^24.0.18",
+    "jest": "^24.9.0",
+    "ts-jest": "^24.1.0"
   },
   "private": true
 }
